+++ internal/wire/datagram_frame.go
@@ -18,7 +18,7 @@ var MaxDatagramSize protocol.ByteCount = 16383
 type DatagramFrame struct {
 	DataLenPresent bool
 	Data           []byte
-	priority       uint8
+	priority       int
 }
 
 func parseDatagramFrame(r *bytes.Reader, typ uint64, _ protocol.Version) (*DatagramFrame, error) {
@@ -90,6 +90,6 @@ func (f *DatagramFrame) GetPriority() uint8 {
 	return f.priority
 }
 
-func (f *DatagramFrame) SetPriority(value uint8) {
+func (f *DatagramFrame) SetPriority(value int) {
 	f.priority = value
 }
